Create SOP 5 : Use of Audio on web pages using HTML5.
Create a webpage named audio.html to set an audio file in webpage with
controls such that it uses HTML5 elements. The audio file must play as soon
as the webpage loads in browser and it will start over again, every time
when it is completed.
Create another webpage named audio1.html which provides multiple source
file formats for the same audio file that plays a sound automatically with
controls. The browser should display the message with appropriate attribute
when audio file is not supported by browser. The code must incorporate the
list of sound files formats (like wav, MP3 or ogg etc).