Create SOP 2 : Create a webpage using HTML and CSS code to design a webpage as the layout displayed below.
The top section will display the heading , ‘Tourist places’ in header. The
section on the left has list of cities. The right hand side display tourist places
in any one of the city .
 Use Inline style sheet in the top section to display background color for the
text ‘Tourist places’. Use internal stylesheet for the left and right section
with background color and font styles.